Update coverage.yaml
In this example, the job runs after the test step and uploads the coverage data to Codecov using the curl command. Again, make sure to replace coverage.xml with the name of your coverage file and rust with the language you're using. You can also customize the authentication tokens and other parameters as needed for your Codecov account.